requirements/docs.txt
sphinx >= 1.5
sphinx_rtd_theme
numpydoc
